+++ scripts/test_media_roundtrip.sh
@@ -19,7 +19,7 @@ cargo build -q -p fluxcast-cli
 ffmpeg -hide_banner -loglevel error -f lavfi -i testsrc=size=160x90:rate=10 -t 1 -c:v libx264 -preset ultrafast -f h264 -y "$task_dir/input.h264"
 ffmpeg -hide_banner -loglevel error -f lavfi -i sine=frequency=440:sample_rate=48000 -t 1 -c:a libopus -y "$task_dir/input.opus"
 
-./target/debug/fluxcast-cli receive-file 127.0.0.1:19110 "$task_dir/output.h264" > "$task_dir/h264.log" 2>&1 & h264_pid=$!
+./target/debug/fluxcast-cli receive-h264 127.0.0.1:19110 > "$task_dir/output.h264" 2> "$task_dir/h264.log" & h264_pid=$!
 ./target/debug/fluxcast-cli receive-file 127.0.0.1:19111 "$task_dir/output.opus" > "$task_dir/opus.log" 2>&1 & opus_pid=$!
 sleep 1
 ./target/debug/fluxcast-cli send-h264 127.0.0.1:19110 "$task_dir/input.h264"